knee-replacement-surgery-evaluation: loco-rs 1.0.1 migration (i64 ids, observability, bg queue)

+++ forms/knee-replacement-surgery-evaluation/back-end-with-loco/src/knee_replacement_surgery_evaluation/models/_entities/knee_replacement_surgery_evaluation_grades.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub oks_total: Option<i32>,
     pub computed_oks_category: String,
@@ -23,7 +23,7 @@ pub struct Model {
     pub signed_by_name: String,
     pub signed_at: Option<DateTimeWithTimeZone>,
     pub graded_at: DateTimeWithTimeZone,
-    pub knee_replacement_surgery_evaluation_id: i32,
+    pub knee_replacement_surgery_evaluation_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
