+++ forms/knee-replacement-surgery-evaluation/back-end-with-loco/src/knee_replacement_surgery_evaluation/models/_entities/knee_replacement_surgery_evaluations.rs
@@ -9,7 +9,7 @@ pub struct Model {
     pub created_at: DateTimeWithTimeZone,
     pub updated_at: DateTimeWithTimeZone,
     #[sea_orm(primary_key)]
-    pub id: i32,
+    pub id: i64,
     pub deleted_at: Option<DateTimeWithTimeZone>,
     pub status: String,
     pub site_name: String,
@@ -95,8 +95,8 @@ pub struct Model {
     pub plan_recommendation: String,
     pub target_list_date: Option<Date>,
     pub responsible_surgeon: String,
-    pub patient_id: i32,
-    pub clinician_id: i32,
+    pub patient_id: i64,
+    pub clinician_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
